bill and ebooklicense are stored float like

+++ src/bookworm/students/rest/StudentResource.java
@@ -78,8 +78,8 @@ public class StudentResource {
         student.setGem(body.getBoolean("isGem"));
         student.setFee(body.getInt("fee"));
         student.setName(body.getString("name"));
-        student.setEBookLicense(body.getInt("eBookLicense"));
-        student.setBill(body.getInt("bill"));
+        student.setEBookLicense(body.getFloat("eBookLicense"));
+        student.setBill(body.getFloat("bill"));
         student.setOwner(getUserId(context));
 
         final boolean success = studentService.save(student);
@@ -142,8 +142,8 @@ public class StudentResource {
         student.setGem(body.getBoolean("isGem"));
         student.setFee(body.getInt("fee"));
         student.setName(body.getString("name"));
-        student.setEBookLicense(body.getInt("eBookLicense"));
-        student.setBill(body.getInt("bill"));
+        student.setEBookLicense(body.getFloat("eBookLicense"));
+        student.setBill(body.getFloat("bill"));
 
         final boolean success = studentService.save(student);
